refactor: 자동차 이름 검증 클래스 디렉토리 변경
