+++ src/main/java/racingcar/domain/car/validator/CarNameValidator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.car;
+package racingcar.domain.car.validator;
 
 import racingcar.constant.errorMessage.ErrorMessages;
 import racingcar.constant.errorMessage.InvalidArgumentException;
